+++ R/pivot-long-SpatVector.R
@@ -30,7 +30,7 @@
 #'
 #' ## `SpatVector`
 #'
-#' The geometry column has a sticky behavior. This means that the result would
+#' The geometry column has a sticky behaviour. This means that the result would
 #' have always the geometry of `data`.
 #'
 #' @examples