+++ tests/testthat/test-arrange-SpatVector.R
@@ -61,3 +61,53 @@ test_that("arrange updates the grouping structure", {
   expect_s4_class(res, "SpatVector")
   expect_equal(as.list(group_rows(res)), list(c(2L, 4L), c(1L, 3L)))
 })
+
+# locale --------------------------------------------------------------
+
+test_that("arrange defaults to the C locale", {
+  skip_on_cran()
+  x <- c("A", "a", "b", "B")
+  df <- tibble(x = x)
+  df$lon <- 1
+  df$lat <- 1
+  df <- terra::vect(df, geom = c("lon", "lat"))
+
+  res <- arrange(df, x)
+  expect_identical(res$x, c("A", "B", "a", "b"))
+
+  res <- arrange(df, dplyr::desc(x))
+  expect_identical(res$x, rev(c("A", "B", "a", "b")))
+})
+
+test_that("locale can be set to an English locale", {
+  skip_on_cran()
+  skip_if_not_installed("stringi", "1.5.3")
+
+  x <- c("A", "a", "b", "B")
+  df <- tibble(x = x)
+  df$lon <- 1
+  df$lat <- 1
+  df <- terra::vect(df, geom = c("lon", "lat"))
+
+  res <- arrange(df, x, .locale = "en")
+  expect_identical(res$x, c("a", "A", "b", "B"))
+})
+
+test_that("non-English locales can be used", {
+  skip_on_cran()
+  skip_if_not_installed("stringi", "1.5.3")
+
+  # Danish `o` with `/` through it sorts after `z` in Danish locale
+  x <- c("o", "\u00F8", "p", "z")
+  df <- tibble(x = x)
+  df$lon <- 1
+  df$lat <- 1
+  df <- terra::vect(df, geom = c("lon", "lat"))
+
+  # American English locale puts it right after `o`
+  res <- arrange(df, x, .locale = "en")
+  expect_identical(res$x, x)
+
+  res <- arrange(df, x, .locale = "da")
+  expect_identical(res$x, x[c(1, 3, 4, 2)])
+})